Consolidate app into main.py
