+++ Stemsplat
@@ -26,4 +26,4 @@ EOF
 fi
 
 cd "${SCRIPT_DIR}"
-python3 install.py
+python3 main.py --install
